Conflicts:
	Statis.Web/ClientBin/Statis.xap
	Statis/Bin/Debug/Statis.dll
	Statis/Bin/Debug/Statis.pdb
	Statis/Bin/Debug/Statis.xap
	Statis/Statis.suo
	Statis/obj/Debug/DesignTimeResolveAssemblyReferences.cache
	Statis/obj/Debug/MainPage.g.cs
	Statis/obj/Debug/MainPage.g.i.cs
	Statis/obj/Debug/Statis.dll
	Statis/obj/Debug/Statis.pdb
	Statis/obj/Debug/Views/Home.g.cs
	Statis/obj/Debug/Views/Home.g.i.cs
	Statis/obj/Debug/XapCacheFile.xml
	StatisServiceHost/bin/Debug/StatisServiceHost.exe
	StatisServiceHost/bin/Debug/StatisServiceHost.pdb
	StatisServiceHost/obj/x86/Debug/ResolveAssemblyReference.cache
	StatisServiceHost/obj/x86/Debug/StatisServiceHost.exe
	StatisServiceHost/obj/x86/Debug/StatisServiceHost.pdb

+++ StatisServiceHost/HandleDb4o.cs
@@ -84,12 +84,22 @@ namespace StatisServiceHost
         }
     
 
-        public static void StoreQuestionnaire(Questionnaire questionnaireToStore)
+        public static void StoreQuestionnaire(string userName, Questionnaire questionnaireToStore)
         {
             var db = Database;
-            DeleteQuestionnaire(db, questionnaireToStore.Name);
-            db.Store(questionnaireToStore);
-            db.Commit();
+            
+            var loggedInUser =
+                (from Analyst user in db
+                 where user.UserName == userName
+                 select user).FirstOrDefault();
+
+            if (loggedInUser != null)
+            {
+                DeleteQuestionnaire(db, questionnaireToStore.Name);
+                loggedInUser.Questionnaires.Add(questionnaireToStore);
+                db.Store(loggedInUser);
+                db.Commit();
+            }
         }
 
         public static void DeleteQuestionnaire(IObjectContainer db, string questionnaireName)
@@ -149,7 +159,7 @@ namespace StatisServiceHost
         public static bool AddAnalyst(string currentUserName, string analystUserName)
         {
             var loggedInUser =
-                (from Analyst user in Database
+                (from Administrator user in Database
                  where user.UserName == currentUserName
                  select user).FirstOrDefault();
